==> 1_the_atlas_core_construction/01_dataimport_preprocessing.rmd <==

# load library

```{r setup, include=FALSE}
library(qs)
library("readxl")
library(Seurat)
library("scCustomize")
library(patchwork)
library(dplyr)
```

# Import and add metadata

## Popovich_2022

```{r setup, include=FALSE}
# Import
setwd("/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Popovich_lab/Single_cell_Faith")
datasetid<-c("SCI00101","SCI00102","SCI00103","SCI00104","SCI00105","SCI00106")
Popovich_2022_dirs <-list.files(".",pattern = "GSM")
Popovich_2022_list <- list()
for (i in seq_along(Popovich_2022_dirs)){
  tmp <- Read10X(
    Popovich_2022_dirs[i],
    unique.features = TRUE,
    strip.suffix = FALSE
  )
  Popovich_2022_list[i] <- CreateSeuratObject(
    tmp,
    assay = "RNA",
    min.cells = 3,
    min.features = 200,
    project = datasetid[i]
  )
}

# Preprocessing

## Annotation: add major cell and microglia annotation using original cell annotations
combine.combined<-qread("faith_final_data.qs")
meta<-combine.combined@meta.data
microglia<-qread("./Faith_original_files/this_micro_obj.qs")
microglia_meta<-microglia@meta.data
rownames(microglia_meta)<-sapply(rownames(microglia_meta), function(x) strsplit(x, split = "-") %>% `[[` (1)%>% head(n = 1))
condition<-unique(meta$orig.ident)
for(i in 1:6){
  tmp<-Popovich_2022_list[[i]]
  first_characters_cellnames<-sapply(colnames(tmp), function(x) strsplit(x, split = "-") %>% `[[` (1)%>% head(n = 1))
  #revise cell name by the way
  tmp<-RenameCells(tmp,new.names  = first_characters_cellnames)
  tmp<-AddMetaData(tmp,paste(condition[i],first_characters_cellnames,sep="_"), col.name = "previous_cellid")
  tmp<-AddMetaData(tmp,meta$celltype[match(tmp$previous_cellid,rownames(meta))],col.name = "previous_celltype")
  tmp$previous_subcelltype<-microglia_meta$subct[match(tmp$previous_cellid,rownames(microglia_meta))]
  Popovich_2022_list[[i]]<-tmp
}

## Change cell id
Popovich_2022_list<-lapply(Popovich_2022_list,function(x){RenameCells(x,add.cell.id = x$orig.ident[1])})

### check cell number
ncol(Popovich_2022_list[[1]])+ncol(Popovich_2022_list[[2]])+ncol(Popovich_2022_list[[3]])+ncol(Popovich_2022_list[[4]])+ncol(Popovich_2022_list[[5]])+ncol(Popovich_2022_list[[6]])
qsave(Popovich_2022_list,"/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Popovich_lab/Single_cell_Faith/Popovich_2022_list.qs")

## QC: Mito genes are not detected in the datasets and it seems mit genes might have been removed by original authors because of 
##technical influences.
Popovich_2022_list[[1]][["percent.mt"]] <-PercentageFeatureSet(Popovich_2022_list[[1]], pattern = "^mt-")
VlnPlot(Popovich_2022_list[[1]], features = c("nFeature_RNA", "nCount_RNA", "percent.mt"), ncol = 3,raster=FALSE)
rownames(Popovich_2022_list[[1]])[grep(rownames(Popovich_2022_list[[1]]),pattern = "^mt-")]

# conclusion: the sham vehicle has 14849 features across 1795 cells.
# notes: check whether we used above to integrate
```

## Lee_2021

```{r setup, include=FALSE}
# import
setwd("/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Lee_lab")
datasetid<-c("SCI00301","SCI00302","SCI00303","SCI00304","SCI00305","SCI00306","SCI00307","SCI00308","SCI00309","SCI00310")
Lee_2021_dirs<-list.files(".",pattern = "txt")
Lee_2021_list<-list()
for (i in seq_along(Lee_2021_dirs)){
  tmp <- read.table(gzfile(Lee_2021_dirs[i]),sep="\t")
  Lee_2021_list[i] <- CreateSeuratObject(
    tmp,
    assay = "RNA",
    min.cells = 3,
    min.features = 200,
    project = datasetid[i]
  )
}

# preprocessing
## add metainfor to the data
lee_meta<-read.table("./GSE162610_barcode_metadata.tsv")
conditions<-c(unique(lee_meta$orig.ident)[8:10],unique(lee_meta$orig.ident)[1:7])
for(i in seq_along(Lee_2021_dirs)){
  tmp<-Lee_2021_list[[i]]@meta.data
  tmp$study_name<-"Lee_2021"
  tmp_meta<-lee_meta[which(lee_meta$orig.ident==conditions[i]),]
  tmp_meta<-tmp_meta[match(rownames(tmp),rownames(tmp_meta)),]
  tmp<-cbind(tmp,tmp_meta)
  colnames(tmp)[33]<-"previous_celltype"
  Lee_2021_list[[i]]@meta.data<-tmp
}

## rename cell id
Lee_2021_list<-lapply(Lee_2021_list,function(x){RenameCells(x,add.cell.id = x$orig.ident[1])})

qsave(Lee_2021_list,"/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Lee_lab/Lee_2021_list.qs")

## QC: don't need to filter cells because it has been filtered and have metadata. More strict can be considered.

# conclusion: we can directly use the first three samples as naive core.
```


## Levine_2022

```{r include=FALSE}
setwd("/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Levine/Matson_Levine_2022")
library(R.utils)
#gunzip("GSE172167_SCI_all_nuclei.rds.gz", remove=FALSE)
GSE172167<-readRDS(file = "./GSE172167_SCI_all_nuclei.rds")
meta_GSE172167<-GSE172167@meta.data
#revise metadata
meta_GSE172167$previous_sampleid<-meta_GSE172167$orig.ident
previous_sampleid<-unique(meta_GSE172167$orig.ident)
sampleid<-c("SCI00401","SCI00402","SCI00403","SCI00404","SCI00405","SCI00406","SCI00407","SCI00408","SCI00409","SCI00410",
            "SCI00411","SCI00412","SCI00413","SCI00414","SCI00415")
  for(i in seq_along(sampleid)){
    meta_GSE172167$orig.ident[which(meta_GSE172167$orig.ident==previous_sampleid[i])]<-sampleid[i]
  }
colnames(meta_GSE172167)[10:12]<-c("previous_subcelltype","previous_celltype","previous_midlevelcelltype")
meta_GSE172167$study_name<-"Levine_2022"
GSE172167@meta.data<-meta_GSE172167
GSE172167<-SplitObject(GSE172167,split.by = "previous_sampleid")
qsave(GSE172167,"/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Levine/Matson_Levine_2022/Levine_2022.qs")
```

## 

```{r}
#---------Khoutorsky_2022
setwd("/fs/ess/PCON0022/guoqi/Integration_spinal_cord/Raw_data_object/Khoutorsky_2022")
naive_id<-c("GSM4960303","GSM4960304","GSM4960305","GSM4960306")
datasetid<-c("SCI01701","SCI01702","SCI01703","SCI01704")
Khoutorsky_2022_list<-list()
for(i in seq_along(naive_id)){
  Khoutorsky_2022_dir<-list.files("./",pattern = naive_id[i])
  matrix<-Read10X_h5(Khoutorsky_2022_dir)
  Khoutorsky_2022_list[[i]] <- CreateSeuratObject(
    matrix,
    assay = "RNA",
    min.cells = 3,
    min.features = 200,
    project = datasetid[i]
  )
}
for(i in seq_along(Khoutorsky_2022_list)){
  meta<-Khoutorsky_2022_list[[i]]@meta.data
  meta$previous_celltype<-"Microglia" #all microglia
  meta$previous_subcelltype<-"Unknown"
  new_meta<-data_sci[match(meta$orig.ident[1],data_sci$`Dataset_ID`),c(2:32)]
  new_meta<-cbind(meta[,1:3],new_meta,meta[,4:5])
  Khoutorsky_2022_list[[i]]@meta.data<-new_meta
  names(Khoutorsky_2022_list)[i]<-as.character(meta$orig.ident[1])
  #print(identical(colnames(new_meta),colnames(meta)))
}
#combine
qsave(Khoutorsky_2022_list,"/fs/ess/PAS1475/guoqi/sci_atlas/raw_data/Khoutorsky_2022/Khoutorsky_2022_filtergene_health_list.qs")
Khoutorsky_2022_list<-qread("/fs/ess/PAS1475/guoqi/sci_atlas/raw_data/Khoutorsky_2022/Khoutorsky_2022_health_list.qs")
health_list_1115<-c(health_list_1115,Khoutorsky_2022_list)
qsave(health_list_1115,"/fs/ess/PAS1475/guoqi/sci_atlas/raw_data/health_list_1115.qs")
```
